Split review screen tests
